Add shared OP level helpers and update references
