+++ bewertung.html
@@ -5,6 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <title>Bewertung</title>
   <link rel="stylesheet" href="interface/ethicom-style.css" />
+  <script src="utils/op-level.js"></script>
   <script src="interface/ethicom-utils.js"></script>
   <script src="interface/accessibility.js"></script>
   <script src="interface/language-selector.js"></script>
