+++ interface/fish-interface/fischEditor.html
@@ -5,6 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <title>Fische Editor</title>
   <link rel="stylesheet" href="../ethicom-style.css" />
+  <script src="../../utils/op-level.js"></script>
   <script src="../ethicom-utils.js"></script>
   <script src="../accessibility.js"></script>
   <script src="../theme-manager.js"></script>
