+++ interface/genealogie.html
@@ -6,6 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <title>Genealogie</title>
   <link rel="stylesheet" href="ethicom-style.css" />
+  <script src="../utils/op-level.js"></script>
   <script src="ethicom-utils.js"></script>
   <script src="accessibility.js"></script>
   <script src="genealogie.js"></script>
