+++ interface/login.html
@@ -6,6 +6,7 @@
   <title>Login</title>
   <link rel="stylesheet" href="ethicom-style.css" />
   <script src="language-selector.js"></script>
+  <script src="../utils/op-level.js"></script>
   <script src="ethicom-utils.js"></script>
   <script src="accessibility.js"></script>
   <script src="login.js"></script>
