+++ interface_OLD/signup.html
@@ -6,6 +6,7 @@
   <title>Registrierung</title>
   <link rel="stylesheet" href="ethicom-style.css" />
   <script src="language-selector.js"></script>
+  <script src="../utils/op-level.js"></script>
   <script src="ethicom-utils.js"></script>
   <script src="signup.js"></script>
   <script src="color-auth.js"></script>
