+++ wings/ratings.html
@@ -6,6 +6,7 @@
   <title>Bewertungen</title>
   <link rel="stylesheet" href="../interface/ethicom-style.css" />
   <link rel="stylesheet" href="wings-style.css" />
+  <script src="../utils/op-level.js"></script>
   <script src="../interface/ethicom-utils.js"></script>
   <script src="../interface/accessibility.js"></script>
   <script src="../interface/language-selector.js"></script>
